requirements.txt
opencv-python>=4.8.0
numpy>=1.24
PyYAML>=6.0
easyocr>=1.7.0
pymupdf>=1.23.0
